requirements.txt
gTTS==2.5.4
moviepy==1.0.3
requests==2.32.3
Pillow==10.4.0
numpy==1.26.4
imageio==2.36.1
imageio-ffmpeg==0.5.1
